refactor(frontend): migrate volumes surface into features
