+++ src/frontend/src/features/layout/sidepanel.tsx
@@ -2,7 +2,7 @@ import { EmptyPanel } from "@/components/patterns/empty-panel";
 import { ErrorBoundary } from "@/components/patterns/error-boundary";
 import { isRlmCoreEnabled, isSectionSupported, UNSUPPORTED_SECTION_REASON } from "@/lib/rlm-api";
 import { useNavigationStore } from "@/stores/navigation-store";
-import { VolumesCanvasPanel } from "@/screens/volumes/volumes-canvas-panel";
+import { VolumesCanvasPanel } from "@/features/volumes/volumes-canvas-panel";
 import {
   WorkspaceCanvasPanel,
   WorkspaceCanvasUnavailablePanel,
