+++ src/frontend/src/features/volumes/__tests__/skill-markdown.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vite-plus/test";
 import { renderToStaticMarkup } from "react-dom/server";
 
-import { SkillMarkdown } from "@/screens/volumes/volumes-canvas-panel";
+import { SkillMarkdown } from "@/features/volumes/volumes-canvas-panel";
 
 describe("SkillMarkdown link sanitization", () => {
   it("renders absolute https links as clickable anchors", () => {
