+++ src/frontend/src/routes/app/volumes.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute, lazyRouteComponent } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/app/volumes")({
-  component: lazyRouteComponent(() => import("@/screens/volumes/volumes-screen"), "VolumesScreen"),
+  component: lazyRouteComponent(() => import("@/features/volumes/volumes-screen"), "VolumesScreen"),
 });
