+++ src/frontend/vite.config.ts
@@ -168,14 +168,14 @@ export default defineConfig({
                 {
                   group: [
                     "@/features/workspace/**",
-                    "@/screens/volumes/**",
+                    "@/features/volumes/**",
                     "!@/features/workspace/workspace-canvas-panel",
                     "!@/features/workspace/workspace-layout-contract",
-                    "!@/screens/volumes/volumes-canvas-panel",
-                    "!@/screens/volumes/volumes-layout-contract",
+                    "!@/features/volumes/volumes-canvas-panel",
+                    "!@/features/volumes/volumes-layout-contract",
                   ],
                   message:
-                    "Layout modules must import screen-owned panels through top-level screen contracts only.",
+                    "Layout modules must import feature-owned panels through top-level feature contracts only.",
                 },
               ],
             },
@@ -223,8 +223,8 @@ export default defineConfig({
         "src/app/workspace/workspace-composer.tsx",
         "src/screens/settings/settings-screen.tsx",
         "src/screens/settings/runtime-form.tsx",
-        "src/screens/volumes/volumes-screen.tsx",
-        "src/screens/volumes/volumes-canvas-panel.tsx",
+        "src/features/volumes/volumes-screen.tsx",
+        "src/features/volumes/volumes-canvas-panel.tsx",
       ],
     },
   },
